feat: set up view context
