+++ src/components/Layout.tsx
@@ -1,7 +1,8 @@
 import Head from 'next/head';
-import React from 'react';
+import React, { useState } from 'react';
 import styled from 'styled-components';
 
+import ViewContext from '../contexts/ViewContext';
 import Nav from './Nav';
 
 const PageWrapper = styled.div`
@@ -19,8 +20,10 @@ const Main = styled.main`
 const Layout: React.FC = function (props) {
   const { children } = props;
 
+  const [view, setView] = useState<string>('colors');
+
   return (
-    <>
+    <ViewContext.Provider value={{ view, setView }}>
       <Head>
         <title>snob</title>
         <meta name="description" content="Generated by create next app" />
@@ -34,7 +37,7 @@ const Layout: React.FC = function (props) {
 
         <footer></footer>
       </PageWrapper>
-    </>
+    </ViewContext.Provider>
   );
 };
