+++ src/contexts/ViewContext.ts
@@ -1,0 +1,10 @@
+import React from 'react';
+
+interface ViewContext {
+  view: string;
+  setView: React.Dispatch<React.SetStateAction<string>>;
+}
+
+const ViewContext = React.createContext<ViewContext | undefined>(undefined);
+
+export default ViewContext;
